requirements.txt
Flask
numpy
tensorflow
Pillow
gunicorn==20.1.0
flask_cors